relocate upstream module index logic to spack.store to avoid cyclic references
